% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{add_terminal_carbons}
\alias{add_terminal_carbons}
\title{add terminal carbons}
\usage{
add_terminal_carbons(dg, zoom = 1)

add_terminal_carbons(dg, zoom = 1)
}
\arguments{
\item{dg}{a Depiction Generator}

\item{a}{Depiction Generator}
}
\description{
add terminal carbons

zoom
}
