% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{add_title}
\alias{add_title}
\title{add_title}
\usage{
add_title(dg)
}
\arguments{
\item{dg}{a CDK  Depiction Generator}
}
\value{
DepictionGenerator
}
\description{
add_title
}
\seealso{
\url{https://cdk.github.io/cdk/latest/docs/api/org/openscience/cdk/depict/DepictionGenerator.html}
}
