% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smarts.R
\name{amino_acids}
\alias{amino_acids}
\title{amino_acids}
\usage{
amino_acids()
}
\description{
return a named list of SMARTS
patterns for amino_acids
}
\seealso{
\url{https://www.daylight.com/dayhtml/doc/theory/theory.smarts.html}

\url{https://www.daylight.com/dayhtml_tutorials/languages/smarts/index.html}
}
