% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{atomcontainer_list_to_jarray}
\alias{atomcontainer_list_to_jarray}
\title{atomcontainer_list_to_jarray}
\usage{
atomcontainer_list_to_jarray(atomcontainers)
}
\value{
a java ArrayList of AtomContainers
}
\description{
atomcontainer_list_to_jarray
}
\examples{
\dontrun {

atmcontnrs <- purrr::map(
     c("CCCCCC", "CCC1CCC1NC", "COCCOCCO")
     parse_smiles)
     
atomcontainer_list_to_jarray(atmcontnrs)

}

}
