% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkJavaClass}
\alias{checkJavaClass}
\title{checkJavaClass}
\usage{
checkJavaClass(jobj, klass)
}
\arguments{
\item{jobj}{Required. A Java Object}

\item{klass}{Required. A string defining a java class}
}
\value{
Boolean
}
\description{
check if an object is of a specified JavaClass
}
