% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{color_atoms}
\alias{color_atoms}
\title{color atoms}
\usage{
color_atoms(dg)
}
\arguments{
\item{dg}{a CDK DepictionGenerator}
}
\value{
DepictionGenerator
}
\description{
set the atom color
}
\seealso{
\url{https://cdk.github.io/cdk/latest/docs/api/org/openscience/cdk/depict/DepictionGenerator.html}
}
