% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{color_background}
\alias{color_background}
\title{color background}
\usage{
color_background(dg, color)
}
\arguments{
\item{dg}{a CDK DepictionGenerator}

\item{color}{a Java Color}
}
\description{
set the background color
}
\examples{
\dontrun{

  dg <- depiction()
  dg |> color_background()

}

}
\seealso{
\url{https://cdk.github.io/cdk/latest/docs/api/org/openscience/cdk/depict/DepictionGenerator.html}
}
