% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict-package.R
\docType{package}
\name{depict-package}
\alias{depict-package}
\title{depict: Draw Beautiful Molecular Structures}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Generate images of molecules using the Depiction API of the 'Chemistry Development Kit'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cdk-r.github.io/depict/}
}

}
\author{
\strong{Maintainer}: Zachary Charlop-Powers \email{zach.charlop.powers@gmail.com} (\href{https://orcid.org/0000-0001-8816-4680}{ORCID})

}
\keyword{internal}
