% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{depict}
\alias{depict}
\title{depict}
\usage{
depict(dg, mol)
}
\arguments{
\item{dg}{Required. A Depiction Generator}

\item{mol}{Required. An AtomContainer}
}
\description{
depict a molecule using the depiction generator.
}
