% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{depiction}
\alias{depiction}
\title{depiction}
\usage{
depiction()
}
\value{
DepictionGenerator
}
\description{
create the CDK depiction object that can be manipulated
}
\seealso{
\url{https://cdk.github.io/cdk/latest/docs/api/org/openscience/cdk/depict/DepictionGenerator.html}
}
