% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{get_image}
\alias{get_image}
\title{get_image}
\usage{
get_image(molgrid)
}
\arguments{
\item{molgrid}{Required. A MolGridDepiction. Usually obtained from
the \code{depict} function.}
}
\description{
get image
}
