% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{get_svg_string}
\alias{get_svg_string}
\title{get_svg_string}
\usage{
get_svg_string(molgrid)
}
\arguments{
\item{molgrid}{Required. A MolGridDepiction. Usually obtained from
the \code{depict} function.}
}
\value{
an SVG string literal
}
\description{
get_svg_string
}
