% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{highlight_atoms}
\alias{highlight_atoms}
\title{highlight_atoms}
\usage{
highlight_atoms(dg, atoms, color)
}
\arguments{
\item{dg}{Required. A Depiction Generator}

\item{atoms}{Required. A list of CDK IChemOjbects, usually atoms or bonds}

\item{color}{Required. A java.awt.Color}
}
\description{
Highlight atoms
}
