% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_functions.R
\name{match_stereo}
\alias{match_stereo}
\title{match_stereo}
\usage{
match_stereo(mol, stereo_type = "any")
}
\arguments{
\item{mol}{Required. an r/CDK AtomContainer}

\item{stereo_type}{Required. a stereo option. One of "None", 
"Bicoordinate", "Tricoordinate", or "Tetracoordinate"}
}
\value{
a hashset of atoms and bonds
}
\description{
return a hashset of Atoms that have the specified stereochemistry
}
