% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{outerglow}
\alias{outerglow}
\title{outerglow}
\usage{
outerglow(dg)
}
\arguments{
\item{dg}{a CDK Depiction Generator}
}
\value{
DepictionGenerator
}
\description{
outerglow
}
\seealso{
\url{https://cdk.github.io/cdk/latest/docs/api/org/openscience/cdk/depict/DepictionGenerator.html}
}
