% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{parse_smiles}
\alias{parse_smiles}
\title{parse_smiles}
\usage{
parse_smiles(smi, generatecoords = TRUE, kekulise = TRUE)
}
\arguments{
\item{smi}{Required. A smiles string}

\item{generatecoords}{Optional. Default \code{TRUE}. Whether
to generate coordinates from the smiles}

\item{kekulise}{Optional. Boolean. Default \code{TRUE}}
}
\value{
an IAtomContainer
}
\description{
simple smiles parser
}
