% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{read_mol}
\alias{read_mol}
\title{read_mol}
\usage{
read_mol(molfile)
}
\arguments{
\item{molfile}{Required. A filepath to a MOLfile.}
}
\value{
an AtomContainer
}
\description{
Read a Molfile and return an AtomContainer
}
