% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{save_image}
\alias{save_image}
\title{save_image}
\usage{
save_image(molgrid, filepath)
}
\arguments{
\item{molgrid}{Required. A MolGridDepiction. Usually obtained from
the \code{depict} function.}

\item{outfile}{Required. Filepath to the output}
}
\description{
Highlight atoms
}
