% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{set_size}
\alias{set_size}
\title{set_size}
\usage{
set_size(dg, width, height)
}
\arguments{
\item{dg}{a CDK Depiction Generator}

\item{width}{width (int)}

\item{height}{height  (int)}
}
\value{
DepictionGenerator
}
\description{
set_size
}
\seealso{
\url{https://cdk.github.io/cdk/latest/docs/api/org/openscience/cdk/depict/DepictionGenerator.html}
}
