% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{set_zoom}
\alias{set_zoom}
\title{set_zoom}
\usage{
set_zoom(dg, zoom = 1)
}
\arguments{
\item{dg}{a Depiction Generator}

\item{zoom}{Optional. Default \code{1}}
}
\value{
DepictionGenerator
}
\description{
set_zoom
}
