% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{add_terminal_carbons}
\alias{add_terminal_carbons}
\title{add terminal carbons}
\usage{
add_terminal_carbons(dg)
}
\arguments{
\item{dg}{a Depiction Generator}
}
\description{
add terminal carbons
}
\seealso{
\url{https://cdk.github.io/cdk/latest/docs/api/org/openscience/cdk/depict/DepictionGenerator.html}
}
