% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{add_terminal_carbons}
\alias{add_terminal_carbons}
\title{add terminal carbons}
\usage{
add_terminal_carbons(dg)
}
\arguments{
\item{dg}{Required. A \href{http://cdk.github.io/cdk/latest/docs/api/org/openscience/cdk/depict/DepictionGenerator.html}{Depiction Generator} probably created by \code{depiction}#' @importFrom  rJava .jcall}
}
\description{
add terminal carbons
}
\examples{

pen  <- parse_smiles("CC1(C(N2C(S1)C(C2=O)NC(=O)CC3=CC=CC=C3)C(=O)[O-])C penicillin")

depiction() \%>\%
  set_size(600, 600) \%>\%
  set_zoom(3) \%>\%
  add_terminal_carbons() \%>\%
  depict(pen) 
  
}
