% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{add_title}
\alias{add_title}
\title{add_title}
\usage{
add_title(dg)
}
\arguments{
\item{dg}{Required. A \href{http://cdk.github.io/cdk/latest/docs/api/org/openscience/cdk/depict/DepictionGenerator.html}{Depiction Generator} probably created by \code{depiction}}
}
\description{
add a title to the depiction. Uses names from the input files.
}
\examples{
pen  <- parse_smiles("CC1(C(N2C(S1)C(C2=O)NC(=O)CC3=CC=CC=C3)C(=O)[O-])C penicillin")

depiction() \%>\%
  set_size(600, 600) \%>\%
  set_zoom(3) \%>\%
  depict(pen) \%>\%
  get_image() \%>\%
  grid::grid.raster()
  
}
