% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{color_atoms}
\alias{color_atoms}
\title{color atoms}
\usage{
color_atoms(dg)
}
\arguments{
\item{dg}{Required. A DepictionGenerator}
}
\description{
color atoms
}
\examples{

pen  <- parse_smiles("CC1(C(N2C(S1)C(C2=O)NC(=O)CC3=CC=CC=C3)C(=O)[O-])C penicillin")

depiction() \%>\%
  set_size(600,600) \%>\%
  set_zoom(3) \%>\%
  color_atoms() \%>\%
  depict(pen) \%>\%
  get_image() \%>\%
  grid::grid.raster()
  

}
