% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict-package.R
\docType{package}
\name{depict-package}
\alias{depict-package}
\alias{_PACKAGE}
\title{Draw Beautiful Molecular Structures}
\description{
Generate images of molecules using the 'Depiction Generator' 
 <http://cdk.github.io/cdk/latest/docs/api/org/openscience/cdk/depict/DepictionGenerator.html> 
 API of  the 'Chemistry Development Kit' <https://cdk.github.io/>.
}
\author{
\strong{Maintainer}: Zachary Charlop-Powers \email{zach.charlop.powers@gmail.com} [copyright holder]

}
\keyword{internal}
