% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{depict}
\alias{depict}
\title{Draw A Molecule and Write it to File}
\usage{
depict(mol, molfile, sizex = NULL, sizey = NULL, type = "png",
  atomcolors = TRUE, moltitle = TRUE, outerglow = TRUE,
  terminalcarbons = TRUE, zoom = 1)
}
\arguments{
\item{mol}{a CDK AtomContainer}
}
\description{
Use CDK's depict moduel to draw a molecule
}
\examples{
mols <- rcdk::parse.smiles("CCC1CCCNNNCCCC1NCN")
mol <- mols[[1]]
depict(mol, molfile="example1.png",zoom=4)

}

