% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{depict}
\alias{depict}
\title{depict}
\usage{
depict(dg, mol)
}
\arguments{
\item{dg}{Required. A DepictionGenerator}

\item{mol}{Required. An AtomContainer}
}
\description{
Depict a molecule using the depiction generator. Display options are set on
the Generator. This function applies those settings to a given molecule or set
of molecules.
}
\seealso{
\url{https://cdk.github.io/cdk/latest/docs/api/org/openscience/cdk/depict/DepictionGenerator.html}

\url{https://cdk.github.io/cdk/latest/docs/api/org/openscience/cdk/AtomContainer.html}
}
