% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{depict}
\alias{depict}
\title{depict}
\usage{
depict(dg, mol)
}
\arguments{
\item{dg}{Required. A Depiction Generator}

\item{mol}{Required. An AtomContainer}
}
\description{
convert a \code{depiction} into an image.
}
\examples{

pen  <- parse_smiles("CC1(C(N2C(S1)C(C2=O)NC(=O)CC3=CC=CC=C3)C(=O)[O-])C penicillin")
dep  <- depiction() 
depict(dep, pen)

}
