% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{get_image}
\alias{get_image}
\title{get_image}
\usage{
get_image(molgrid)
}
\arguments{
\item{molgrid}{Required. A MolGridDepiction. Usually obtained from
the \code{depict} function.}
}
\description{
get an image generated in Java as an R-accessible image.
}
\examples{

pen      <- parse_smiles("CC1(C(N2C(S1)C(C2=O)NC(=O)CC3=CC=CC=C3)C(=O)[O-])C penicillin")
depicted <- depiction() \%>\% depict(pen)
image    <- get_image(depicted)
\dontrun{
grid::grid.raster(image)
}
  
}
