% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{highlight_atoms}
\alias{highlight_atoms}
\title{highlight_atoms}
\usage{
highlight_atoms(dg, atoms, color)
}
\arguments{
\item{dg}{Required. A \href{http://cdk.github.io/cdk/latest/docs/api/org/openscience/cdk/depict/DepictionGenerator.html}{Depiction Generator} probably created by \code{depiction}}

\item{atoms}{Required. CDK atoms or bond. Possible specified by \code{match_smarts}}

\item{color}{Required. A color of Java class \code{java.awt.Color}}
}
\description{
Highlight atoms in a depiction.
}
\examples{

pen    <- parse_smiles("CC1(C(N2C(S1)C(C2=O)NC(=O)CC3=CC=CC=C3)C(=O)[O-])C penicillin")
lactam <- match_smarts("C1(=O)NCC1", pen)
color  <- J("java.awt.Color")

depiction() \%>\%
  highlight_atoms(lactam, color$RED) \%>\%
  depict(pen)
  
}
