% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_functions.R
\name{match_smarts}
\alias{match_smarts}
\title{match_smarts}
\usage{
match_smarts(smarts, mol, limit = 10)
}
\arguments{
\item{smarts}{a SMARTS string}

\item{mol}{a CDK IAtomContainer or a java.util.Arraylist of them}

\item{limit}{limit of the number of matches}
}
\value{
a hashset of atoms and bonds
}
\description{
Find matches to a given molecule using a SMARTs pattern.
}
\seealso{
\url{https://cdk.github.io/cdk/latest/docs/api/org/openscience/cdk/isomorphism/Mappings.html}

\url{https://cdk.github.io/cdk/latest/docs/api/org/openscience/cdk/smiles/smarts/SmartsPattern.html}
}
