% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_functions.R
\name{match_smarts}
\alias{match_smarts}
\title{match SMARTs}
\usage{
match_smarts(smarts, mol, limit = 10)
}
\arguments{
\item{smarts}{Required. a smarts string}

\item{mol}{Required. an r/CDK AtomContainer}

\item{limit}{Optional. Default \code{10}}
}
\value{
a hashset of atoms and bonds
}
\description{
match SMARTs
}
