% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{save_image}
\alias{save_image}
\title{save_image}
\usage{
save_image(molgrid, outfile)
}
\arguments{
\item{molgrid}{Required. A \code{MolGridDepiction}. Usually obtained from
the \code{depict} function.}

\item{outfile}{Required. A file path.}
}
\description{
save the output of the \code{depict} function.
}
\examples{
pen    <- parse_smiles("CC1(C(N2C(S1)C(C2=O)NC(=O)CC3=CC=CC=C3)C(=O)[O-])C penicillin")
lactam <- match_smarts("C1(=O)NCC1", pen)
color  <- J("java.awt.Color")

depiction() \%>\%
  highlight_atoms(lactam, color$LIGHT_GRAY) \%>\%
  depict(pen) \%>\%
  save_image("penicillin.png")
  
}
