% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{set_size}
\alias{set_size}
\title{set_size}
\usage{
set_size(dg, width, height)
}
\arguments{
\item{dg}{Required. a Depiction Generator}

\item{width}{Required. width}

\item{height}{Required. height}
}
\description{
set_size
}
