% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depict.R
\name{set_size}
\alias{set_size}
\title{set_size}
\usage{
set_size(dg, width, height)
}
\arguments{
\item{dg}{Required. a Depiction Generator}

\item{width}{Required. width}

\item{height}{Required. height}
}
\description{
set canvas size for depictiongenerator
}
\examples{

pen  <- parse_smiles("CC1(C(N2C(S1)C(C2=O)NC(=O)CC3=CC=CC=C3)C(=O)[O-])C penicillin")

depiction() \%>\%
  set_size(600, 600) \%>\%
  set_zoom(3) \%>\%
  depict(pen) \%>\%
  get_image() \%>\%
  grid::grid.raster()

}
